pyproject.toml
[tool.poetry]
name = "django-payments-paybox"
version = "0.1.0"
description = "django-payments provider for PayBox.money"
authors = ["vlzh <vlzh@pm.me>"]

[tool.poetry.dependencies]
python = "^3.6"
pydantic = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
django-payments = "^0.13.0"
django = "^3.0.8"
flake8-black = "^0.2.1"
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
